This file allows a developer to add a new tool in either the C or Python trees, add the executable (and any supporting files, such as the TestSigningPrivateKey.pem file) to this file and the build server will automatically make sure that the new files are added to the BaseTools\Bin\Win32 directory. The Win32 directory is located in https://svn.code.sf.net/p/edk2-toolbinaries/code/trunk/Win32 repository.
Developer - Tool add process:
1)	Developer adds code for the new tool.
2)	Developer updates the Makefile in the C or Python directory
a)	The entry must make sure that the executable is generated in the BaseTools\Bin\Win32 directory and any supporting files are copied to the same directory as part of the build step.
3)	Developer adds the <Toolname>.exe under the [Bin.Win32] section in the BinaryFiles.txt file.
4)	Developer adds other files required to be present in the [ExtraFiles.Win32] section in the BinaryFiles.txt file.

Build Server:
1)	Build all binaries by calling nmake on the Source\C\Makefile and Source\Python\Makefile
2)	After building the binaries, the build server verify that the files listed in BaseTools\Source\BinFiles.txt are also in the edk2-toolbinaries project,
a.	If a file is not under source control, then the build server will add file as long as it is present.

File format:
[SectionName.TargetDir]
File1
File2
…

Where: 
SectionName is one of Bin, ExtraFiles or CxFreeze
TargeDir is the name of the subdirectory in the BaseTools\Bin directory tree.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: lhauch <larry.hauch@intel.com>


git-svn-id: svn://svn.code.sf.net/p/edk2/code/trunk/edk2/BaseTools@16063 6f19259b-4bc3-4df7-8a09-765794883524